add Icon.js, Navbar.js to make burger menu work
